+++ lightly_studio_view/src/lib/components/QueryEditor/useLightlyQueryLanguage/syntaxDocumentation/buildFieldHover.ts
@@ -1,10 +1,10 @@
 import * as monaco from 'monaco-editor';
-import { SCOPES, type FieldDoc, type Scope } from '../../language/lightly-query-schema';
+import { SCOPES } from '../../language/lightly-query-schema';
 
 /** Build hover content for a field in a resolved scope, including type and description. */
-export function buildFieldHover(
-    scope: Scope,
-    field: FieldDoc,
+export function buildFieldHover<S extends keyof typeof SCOPES>(
+    scope: S,
+    field: (typeof SCOPES)[S]['fields'][number],
     range: monaco.Range
 ): monaco.languages.Hover {
     const { title } = SCOPES[scope];
